requirements.txt
numpy
pandas
tensorflow
paho-mqtt
joblib
requests
